cf_scorecheck/requirements.txt
beautifulsoup4==4.12.3
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
firebase-admin==6.4.0
google==3.0.0
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-cloud-container==2.42.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-secret-manager==2.18.3
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.0
grpcio-status==1.62.0
httplib2==0.22.0
idna==3.6
kubernetes==29.0.0
msgpack==1.0.8
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
uritemplate==4.1.1
urllib3==2.2.1
websocket-client==1.7.0